Added CommonModuleResolverTrait
